Finnished pretty big refactor, client split into three parts: client, session and connection
