--- Sources/SwiftMQTT/MQTT/MQTTClient/MQTTClientnew.swift
@@ -1,173 +1,0 @@
-actor MQTTClientnew {
-    private let config: Config
-    private let clientId: String
-
-    private let internalEventBus: MQTTEventBus<MQTTInternalEvent>
-    private let eventBus: MQTTEventBus<MQTTEvent>
-    private let internalCommandBus: MQTTEventBus<MQTTInternalCommand>
-
-    private let internalEventStream: AsyncStream<MQTTInternalEvent>
-    private let internalCommandStream: AsyncStream<MQTTInternalCommand>
-    let eventStream: AsyncStream<MQTTEvent>
-
-
-    private let session: MQTTSession
-    private let connection: MQTTConnection
-    private let idAllocator: PacketIdAllocator
-
-    private var keepAliveTask: Task<Void, Never>?
-
-    init(clientId: String, host: String, port: Int, config: Config) {
-        self.clientId = clientId
-        self.config = config
-
-        var internalCont: AsyncStream<MQTTInternalEvent>.Continuation!
-        var internalCommandCont: AsyncStream<MQTTInternalCommand>.Continuation!
-        var cont: AsyncStream<MQTTEvent>.Continuation!
-
-        // TODO: Should probably let user define how many events to buffer
-        self.internalEventStream = AsyncStream(bufferingPolicy: .bufferingNewest(10)) { internalCont = $0 }
-        self.eventStream = AsyncStream(bufferingPolicy: .bufferingNewest(10)) { cont = $0 }
-
-        self.internalEventBus = MQTTEventBus<MQTTInternalEvent>(continuation: internalCont)
-        self.eventBus = MQTTEventBus<MQTTEvent>(continuation: cont)
-
-        self.internalCommandStream = AsyncStream(bufferingPolicy: .bufferingNewest(10)) { internalCommandCont = $0 }
-        self.internalCommandBus = MQTTEventBus<MQTTInternalCommand>(continuation: internalCommandCont )
-
-        self.session = MQTTSession(config: config, eventBus: eventBus, commandBus: internalCommandBus)
-        self.connection = MQTTConnection(host: host, port: port, eventBus: internalEventBus)
-
-        self.idAllocator = .init()
-
-        // Read events from connection
-        Task {
-            for await event in self.internalEventStream {
-                await self.session.handle(event)
-            }
-        }
-
-        // Read commands from session
-        Task {
-            for await command in self.internalCommandStream {
-                switch command {
-                    case .send(let packet):
-                        // should handle error here
-                        try? await self.send(packet)
-                }
-            }
-        }
-    }
-}
-
-extension MQTTClientnew {
-    func connect() async throws {
-        try await self.tryConnect()
-        self.startKeepAlive()
-    }
-
-    private func tryConnect() async throws {
-        try await self.connection.send(packet: Connect(clientId: self.clientId, keepAlive: 60))
-        try await self.session.awaitConack()
-    }
-
-    private func reconnect() async throws {
-        var attempts = 0
-
-        while attempts <= self.config.maxRetries {
-            do {
-                try await self.tryConnect()
-                break
-            } catch {
-                attempts += 1
-            }
-        }
-    }
-}
-
-// MARK: Send
-extension MQTTClientnew {
-    private func send(_ packet: any MQTTControlPacket) async throws {
-        await session.handle(.send(packet))
-        try await self.connection.send(packet: packet)
-    }
-}
-
-// MARK: Publish
-extension MQTTClientnew {
-    func publish(bytes: Bytes, qos: QoS, topic: String) async throws {
-        let packetId = await idAllocator.next()
-        let publish = Publish(topicName: topic, message: bytes, packetId: packetId, qos: qos)
-
-        try await sendPublish(publish)
-
-        switch qos {
-            case .ExactlyOnce:
-                try await self.session.awaitPubrec(packetId: packetId)
-                try await self.send(Pubrec(packetId: packetId))
-                try await self.session.awaitPubComp(packetId: packetId)
-            case .AtLeastOnce:
-                try await session.awaitPuback(packetId: packetId)
-            case .AtMostOnce:
-                break
-        }
-    }
-
-    func publish(message: String, qos: QoS, topic: String) async throws {
-        let packetId = await idAllocator.next()
-        let publish = Publish(topicName: topic, message: message, packetId: packetId, qos: qos)
-
-        try await sendPublish(publish)
-    }
-
-    private func sendPublish(_ publish: Publish) async throws {
-        try await self.send(publish)
-    }
-}
-
-// MARK: Subscribe
-extension MQTTClientnew {
-    func subscribe(to topics: [TopicFilter]) async throws {
-            let packetId = await self.idAllocator.next()
-            let subscribePacket = Subscribe(packetId: packetId, topics: topics)
-
-            try await self.connection.send(packet: subscribePacket)
-            try await self.session.awaitSuback(packetId: packetId)
-    }
-
-    private func subscribeToTopics() async throws {
-        let topics = await self.session.getSubscriptions()
-        if topics.count <= 0 { return }
-
-        try await self.subscribe(to: topics)
-    }
-}
-
-// MARK: Keepalive
-extension MQTTClientnew {
-    private func startKeepAlive() {
-        let task = Task {
-            while !Task.isCancelled {
-                await session.awaitKeepAlive()
-
-                // TODO: handle error
-                try? await self.send(Pingreq())
-                // try session.awaitPingResp()
-                do {
-                    try await session.awaitPingresp()
-                } catch {
-                    // reconnect
-                    do {
-                        self.eventBus.emit(.error(error))
-                        try await self.reconnect()
-                    } catch {
-                        // TODO: call disconnect(with: error)
-                        self.eventBus.emit(.error(error))
-                    }
-                }
-            }
-        }
-
-        self.keepAliveTask = task
-    }
-}
